Remove version from implementation (#165)
* Backend: expand path with version

* remove version handling

* remove version handling

* TST: fix test

* TST: fix test of docstring examples

* DOC: fix SQL example

* fix linter errors

* TST: fix use backend path separator when expanding path

* Backend: check path and version in _path_with_version()

* Backend.latest_version(): remove path check

* Backend._ls(): improve docstring
